Version 2.6.1 (#7)

+++ include/graphtyper/graph/absolute_position.hpp
@@ -15,8 +15,8 @@ struct Contig;
 class AbsolutePosition
 {
 public:
-  std::vector<uint32_t> offsets;
-  std::unordered_map<std::string, uint32_t> chromosome_to_offset;
+  std::vector<long> offsets;
+  std::unordered_map<std::string, long> chromosome_to_offset;
 
   AbsolutePosition() = default;
   AbsolutePosition(std::vector<Contig> const & contigs);
@@ -26,10 +26,10 @@ public:
 
   ///* const methods */
   bool is_contig_available(std::string const & chromosome) const;
-  uint32_t get_absolute_position(std::string const & chromosome, uint32_t contig_position) const;
+  long get_absolute_position(std::string const & chromosome, long contig_position) const;
 
-  std::pair<std::string, uint32_t>
-  get_contig_position(uint32_t absolute_position,
+  std::pair<std::string, long>
+  get_contig_position(long absolute_position,
                       std::vector<Contig> const & contigs) const;
